fixed some testcases

git-svn-id: http://snes-sdk.googlecode.com/svn/trunk@55 5e493d80-052e-11de-a3aa-254263d3d64b

+++ tcc-65816/test/unimplemented/dg/math/builtins-24.c
@@ -9,15 +9,15 @@
 /* { dg-do run } */
 /* { dg-options "-O2 -ffast-math" } */
 
-extern double pow(double,double);
+extern float pow(float,float);
 extern void abort(void);
 
-double foo (double x)
+float foo (float x)
 {
   return pow (x, 6);
 }
 
-double bar (double x)
+float bar (float x)
 {
   return pow (x, -4);
 }
